requirement.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-multipart==0.0.9
jinja2==3.1.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.13.3
python-dotenv==1.0.1